[IK-341] refactor:  프로젝트 정리
